ClaudeCode/requirements.txt
# Excel I/O and manipulation
openpyxl>=3.1.2
pandas>=2.0.0
numpy>=1.24.0

# Formula parsing and evaluation
formulas>=1.2.0

# Testing and validation
pytest>=7.4.0
pytest-xdist>=3.3.1
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Code quality
black>=23.0.0
pylint>=2.17.0
mypy>=1.4.0
